routes for all rewards and create rewards

+++ app/forms/reward_form.py
@@ -1,9 +1,9 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, DecimalField
+from wtforms import StringField, DecimalField, IntegerField
 from wtforms.validators import DataRequired
 
 class RewardForm(FlaskForm):
     title = StringField("Reward Title", validators=[DataRequired(message="Reward title cannot be blank.")])
-    price = DecimalField("Price", validators=[DataRequired(message="Price for this reward must be set.")])
+    price = IntegerField("Price", validators=[DataRequired(message="Price for this reward must be set.")])
     description = StringField("Description", validators=[DataRequired(message="Reward description cannot be blank.")])
     estimatedDelivery = StringField("Estimated Delivery", validators=[DataRequired(message="Reward estimated delivery cannot be blank.")])
